+++ client/app/bundles/course/reference-timelines/TimelineDesigner.tsx
@@ -1,10 +1,10 @@
 import LoadingIndicator from 'lib/components/core/LoadingIndicator';
 import Preload from 'lib/components/wrappers/Preload';
+import { useAppDispatch } from 'lib/hooks/store';
 
 import DayView from './views/DayView';
 import { LastSavedProvider } from './contexts';
 import { fetchTimelines } from './operations';
-import { useAppDispatch } from './store';
 
 const TimelineDesigner = (): JSX.Element => {
   const dispatch = useAppDispatch();
