+++ client/app/bundles/course/reference-timelines/store/index.ts
@@ -1,4 +1,3 @@
-export * from './hooks';
-export { default as store } from './store';
 export * from './selectors';
+export { default as store } from './store';
 export { timelinesActions as actions } from './timelinesReducer';
